+++ apps/studio/src/background/NativeMenuActionHandlers.ts
@@ -4,10 +4,10 @@ import { buildWindow, getActiveWindows, OpenOptions } from './WindowBuilder'
 import { app , shell } from 'electron'
 import platformInfo from '../common/platform_info'
 import path from 'path'
-import { SavedConnection } from '../common/appdb/models/saved_connection'
 import { IGroupedUserSettings } from '../common/appdb/models/user_setting'
 import { IMenuActionHandler } from '@/common/interfaces/IMenuActionHandler'
 import { autoUpdater } from "electron-updater"
+import { SavedConnection } from '@/lib/utility/appdb/SavedConnection'
 
 type ElectronWindow = Electron.BrowserWindow | undefined
